+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
 ### Fixed
 
-### Changed
+## [sdr_v2.1.2.7] - 2023-05-31
+
+### Fixed
+
+- In dust and sea-salt, changed dimensions back to `globalCellCountPerDim` since these are needed to determine emission tuning parameters, not to allocate arrays.
 
 ## [sdr_v2.1.2.6] - 2023-05-06